HSEARCH-3089 Move field predicate builder factory type errors to a common abstract class
Just to avoid copy-pasted code.
